duration icalendar converters
